lat/long 5 dec, user chooses shortID
The admin is now allowed to choose their own ShortID. The binary has
been updated so that only ShortIDs that are multiplies of 5 can be used.
This allows the other 4 GCAs to use IDs that are multiples of 5 mod 1, 5
mod 2, etc... this ensures that multiple GCAs creating boxes together
will not unintentionally use the same ShortID for a single piece of
hardware.

The lat/long was expanded to 5 decimals because it turns out that 3
decimals only gives you a resolution to about 100 meters, which isn't
good enough for smaller solar installations.